Implemented single connection setup. With this commit, the database connection is performed once, and the same connection is used for sub-operations.
This makes it much cleaner when a connection is passed to a sub-operation, and allows the backend manager to directly write log data to the database without needing to pass the connection around.

For future updates, this also makes it possible to move the transaction into the manager, instead of having to pass around the transaction as it is done now.