Fixed code to use explicit logging around database operations in backup mode.
Abstracted all database interactions into custom classes so there is now locking around reads and writes to the database, so we better observe the rules for SQLite.

Updated last operations to be async and correctly use the cancellation tokens.

Fixed use of transactions to mimic the SQLite way of only having a single transaction per connection, and no nested transactions.