PureFTP RecursionLimit detection and Warnings
This addresses the issue described on
https://github.com/duplicati/duplicati/issues/6005

If a PureFTP server is detected, a warning is issued when executing server file listing.

The parameter`ignore-pureftpd-limit-issue` allows the warning to be suppressed.

In the event of an effective truncation, it will throw an exception and that cannot be suppressed as incomplete listing can lead to misleading backup/restore operations.