Warn more on extra files
If there are extra unknown files, this can be a sign that multiple backups are using the same folder.

It can also be a sign that an older database is used for some reason.

This PR updates the error message to mention these two common cases, and spell out what the repair process will do.