Fixed timestamp drift causing validation errors
If the upload of the dlist fails, the filename will be incremented with one second. This causes validation errors when running repair, because it compares the fileset times (when the backup actually started) with the timestamp in the filename.

This PR updates the check to look only at the filename timestamp so the comparison is accurate even in the failure scenario.

This fixes #6235