Apply Secret Provider to TrayIcon
If the secret provider is used to supply arguments only used by the TrayIcon it does not work.

The secret provider is applied in the Server part, so any settings passed through are correctly resolved, but for the few options that apply directly to the TrayIcon this was not resolved.

With this PR, the secret provider is applied to TrayIcon options as well.

This applies to:
- `--hosturl`
- `--browser-command`
- `--host-cert-hash`
- `--webservice-password`