Run backups in order after startup
This changes the order that backups are queue to better match their intended execution order.

If the machine is off for a day, the next time Duplicati starts, it will find all backups that were supposed to have run and queues these.

Before, the order of backups being queued would depend on the database order, essentially creation order.

This PR changes the logic to keep track of when they were supposed to run (first triggered time if multiple slots were skipped), and then adds the backups into the queue, so the job that was supposed to run first, will now be queued first.

This fixes #1476