Faster progress update
The default is to report progress every second.
In some cases the report output is stuck at the "Starting backup..." phase.

This PR adds a new `Starting_Operation` enum to indicate that we do not yet know which operation is starting.

Then an additional trigger is added to flush the progress when the pahse changes. This means that as soon as we know which operation is active, this will be flushed to the UI.