Update Docker images
This updates Docker images to support UID/GID overrides, faling back to running as root.

This also fixes the images to be actually allow executing the binaries as described in the README by adding them to the PATH environment variable.

The README file was overhauled to fix some outdated content and include the new UID/GID override feature.