Added more guards on MSGraph errors
This PR handles empty responses with the status code 204 - No Content.

Prior to this PR a 204 response when cancelling an upload session would cause an error due to not being able to parse the empty body.

This should get us one step closer to figuring out what is triggering the errors.