Fix timeout issue on MSGraph
This fixes the reported timeout issue with MicrosoftGraphBackend cause by each fragment being required to be uploaded within the `--short-timeout` value.

With this PR the overall progress is monitored with `--read-write-timeout` (as before) and each fragment's speed is not measured.

Also updated the fragment numbers with +1 so the first fragment is "1/5" instead of "0/5", such that the last fragment is "5/5".