Show assembly version
This PR fixes a long-standing issue where the reported version number is just the tag with a `-` prefix.

This was caused by the logic dropping the assembly version number if a tag exists, but prefixing it with a dash.

The fix is to now show both the assembly version and the tag.