Added NOSCRIPT feature to MSI
This PR adds an MSI property that can be set with `NOSCRIPT=true`.

Setting the property will disable scripting in the MSI and thus prevent starting/stopping running services.

Also made the stop-service script non-critial, so it is possible to install the MSI even if the script fails to start or returns an error.