Added support for Synology DSM
This adds back support for Synology DSM.

The implementation is compatible with DSM 7.2+ and integrates with DSM to forward requests through the DSM server into Duplicati.

The setup features database locks, using a random auth-key so DSM handles authentication.

The database is encrypted with a random key, and all data is stored in the `var` folder.

The uninstall can optionally remove the `var` data.

This is still a work in progress. The authentication integration with DSM does not currently work.